Controllers atualizados

+++ ApiCatalogoJogos/Controllers/v1/UsuarioController.cs
@@ -3,8 +3,11 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
+using ApiCatalogoJogos.Business.Entities.Named;
 using ApiCatalogoJogos.Business.Exceptions;
 using ApiCatalogoJogos.Business.Services;
+using ApiCatalogoJogos.Enum;
+using ApiCatalogoJogos.Infrastructure.Authorization;
 using ApiCatalogoJogos.Infrastructure.Model.InputModel;
 using ApiCatalogoJogos.Infrastructure.Model.ViewModel;
 using Microsoft.AspNetCore.Mvc;
@@ -12,6 +15,7 @@ using Swashbuckle.AspNetCore.Annotations;
 
 namespace ApiCatalogoJogos.Controllers.v1
 {
+    [Authorize(PermissaoUsuario.Usuario, PermissaoUsuario.Administrador)]
     [Route("api/v1/usuarios")]
     [ApiController]
     public class UsuarioController : ControllerBase
@@ -30,7 +34,9 @@ namespace ApiCatalogoJogos.Controllers.v1
         /// <param name="quantidade">Define a quantidade em cada pagina</param>
         [SwaggerResponse(statusCode: 200, description: "Retorna usuários recuperados", Type = typeof(List<UsuarioViewModel>))]
         [SwaggerResponse(statusCode: 204, description: "Nenhum usuário na página")]
+        [SwaggerResponse(statusCode: 401, description: "Permissão insuficiente")]
         [SwaggerResponse(statusCode: 500, description: "Erro interno")]
+        [Authorize(PermissaoUsuario.Administrador)]
         [HttpGet]
         public async Task<ActionResult<IEnumerable<UsuarioViewModel>>> Obter(
             [FromQuery, Range(1, int.MaxValue)] int pagina = 1,
@@ -49,6 +55,7 @@ namespace ApiCatalogoJogos.Controllers.v1
         /// </summary>
         /// <param name="id">Id do usuário</param>
         [SwaggerResponse(statusCode: 200, description: "Retorna o usuário com id informado", Type = typeof(UsuarioViewModel))]
+        [SwaggerResponse(statusCode: 401, description: "Permissão insuficiente")]
         [SwaggerResponse(statusCode: 404, description: "Usuário não encontrado")]
         [SwaggerResponse(statusCode: 500, description: "Erro interno")]
         [HttpGet("{id:guid}")]
@@ -56,6 +63,9 @@ namespace ApiCatalogoJogos.Controllers.v1
         {
             try
             {
+                if (!ValidaSessao(id))
+                    return Unauthorized();
+
                 var usuario = await _service.Obter(id);
                 return Ok(usuario);
             }
@@ -73,6 +83,7 @@ namespace ApiCatalogoJogos.Controllers.v1
         [SwaggerResponse(statusCode: 400, description: "Erro nos dados informados")]
         [SwaggerResponse(statusCode: 422, description: "Erro durante a inserção")]
         [SwaggerResponse(statusCode: 500, description: "Erro interno")]
+        [AllowAnonymous]
         [HttpPost]
         public async Task<ActionResult<JogoViewModel>> Inserir([FromBody] UsuarioInputModel usuarioInput)
         {
@@ -98,6 +109,7 @@ namespace ApiCatalogoJogos.Controllers.v1
         /// <param name="id">Id do usuário a ser atualizado</param>
         /// <param name="usuarioInput">Usuário com as novas características configuradas</param>
         [SwaggerResponse(statusCode: 200, description: "Retorna o usuário atualizado", Type = typeof(UsuarioViewModel))]
+        [SwaggerResponse(statusCode: 401, description: "Permissão insuficiente")]
         [SwaggerResponse(statusCode: 400, description: "Erro nos dados informados")]
         [SwaggerResponse(statusCode: 404, description: "Usuário não encontrado")]
         [SwaggerResponse(statusCode: 500, description: "Erro interno")]
@@ -106,6 +118,9 @@ namespace ApiCatalogoJogos.Controllers.v1
         {
             try
             {
+                if (!ValidaSessao(id))
+                    return Unauthorized();
+
                 var usuario = await _service.Atualizar(id, usuarioInput);
 
                 return Ok(usuario);
@@ -122,6 +137,7 @@ namespace ApiCatalogoJogos.Controllers.v1
         /// <param name="id">Id do usuário a ser atualizado</param>
         /// <param name="quant">Nova quantia de fundos</param>
         [SwaggerResponse(statusCode: 200, description: "Retorna o usuário atualizado", Type = typeof(UsuarioViewModel))]
+        [SwaggerResponse(statusCode: 401, description: "Permissão insuficiente")]
         [SwaggerResponse(statusCode: 400, description: "Erro nos dados informados")]
         [SwaggerResponse(statusCode: 404, description: "Usuário não encontrado")]
         [SwaggerResponse(statusCode: 500, description: "Erro interno")]
@@ -130,6 +146,9 @@ namespace ApiCatalogoJogos.Controllers.v1
         {
             try
             {
+                if (!ValidaSessao(id))
+                    return Unauthorized();
+
                 var usuario = await _service.AtualizarFundos(id, quant);
 
                 return Ok(usuario);
@@ -146,6 +165,7 @@ namespace ApiCatalogoJogos.Controllers.v1
         /// <param name="id">Id do usuário</param>
         /// <param name="idJogo">Id do jogo a ser adicionado</param>
         [SwaggerResponse(statusCode: 200, description: "Retorna o usuário atualizado", Type = typeof(UsuarioViewModel))]
+        [SwaggerResponse(statusCode: 401, description: "Permissão insuficiente")]
         [SwaggerResponse(statusCode: 400, description: "Erro nos dados informados")]
         [SwaggerResponse(statusCode: 404, description: "Usuário não encontrado")]
         [SwaggerResponse(statusCode: 500, description: "Erro interno")]
@@ -154,6 +174,9 @@ namespace ApiCatalogoJogos.Controllers.v1
         {
             try
             {
+                if (!ValidaSessao(id))
+                    return Unauthorized();
+
                 var usuario = await _service.AdicionarJogo(id, idJogo);
 
                 return Ok(usuario);
@@ -169,6 +192,7 @@ namespace ApiCatalogoJogos.Controllers.v1
         /// </summary>
         /// <param name="id">Id do usuário a ser removido</param>
         [SwaggerResponse(statusCode: 200, description: "Retorna o id do usuário removido")]
+        [SwaggerResponse(statusCode: 401, description: "Permissão insuficiente")]
         [SwaggerResponse(statusCode: 404, description: "Usuário não encontrado")]
         [SwaggerResponse(statusCode: 500, description: "Erro interno")]
         [HttpDelete("{id:guid}")]
@@ -176,6 +200,9 @@ namespace ApiCatalogoJogos.Controllers.v1
         {
             try
             {
+                if (!ValidaSessao(id))
+                    return Unauthorized();
+
                 await _service.Remover(id);
 
                 return Ok(id);
@@ -185,6 +212,35 @@ namespace ApiCatalogoJogos.Controllers.v1
                 return NotFound(ex.Message);
             }
         }
+
+        /// <summary>
+        /// Autentica usuário cadastrado
+        /// </summary>
+        /// <param name="inputModel">InputModel contendo dados para login</param>
+        [SwaggerResponse(statusCode: 200, description: "Retorna o token de autenticação e ViewModel do usário autenticado")]
+        [SwaggerResponse(statusCode: 404, description: "Usuário não encontrado")]
+        [SwaggerResponse(statusCode: 500, description: "Erro interno")]
+        [AllowAnonymous]
+        [HttpPost("login")]
+        public async Task<ActionResult> Autenticar(LoginInputModel inputModel)
+        {
+            try
+            {
+                var auth = await _service.Autenticar(inputModel);
+                return Ok(new JsonResult(new {token = auth.Item1, usuario = auth.Item2 }).Value);
+            }
+            catch (AutorizacaoException ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+        
+        // Util
+        private bool ValidaSessao(Guid id)
+        {
+            var usuario = (UsuarioViewModel)HttpContext.Items["Usuario"];
+
+            return (id == usuario.Id || usuario.Permissao == PermissaoUsuario.Administrador);
+        }
     }
 }
-
